+++ framework/packages/abstract-sdk/src/apis/stargate/feegrant.rs
@@ -1,3 +1,4 @@
+use cosmos_sdk_proto::cosmos::{base, feegrant};
 use cosmwasm_std::{Coin, Timestamp};
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
@@ -86,5 +87,60 @@ impl<A: BasicOrPeriodicAllowance> AllowedMsgAllowance<A> {
 /// Trait for types that can be used as allowances in the FeeGranter.
 pub trait BasicOrPeriodicAllowance: MsgAllowance {}
 
+impl BasicOrPeriodicAllowance for BasicAllowance {}
+impl BasicOrPeriodicAllowance for PeriodicAllowance {}
+
 /// Trait for types that can be used as feegrant type
 pub trait MsgAllowance: StargateMessage {}
+
+impl MsgAllowance for BasicAllowance {}
+impl MsgAllowance for PeriodicAllowance {}
+impl<A: BasicOrPeriodicAllowance> MsgAllowance for AllowedMsgAllowance<A> {}
+
+// Stargate Msg implementations
+
+impl StargateMessage for BasicAllowance {
+    type ProtoType = feegrant::v1beta1::BasicAllowance;
+
+    fn to_proto(&self) -> feegrant::v1beta1::BasicAllowance {
+        feegrant::v1beta1::BasicAllowance {
+            spend_limit: self
+                .spend_limit
+                .iter()
+                .map(|item| base::v1beta1::Coin {
+                    denom: item.denom.clone(),
+                    amount: item.amount.to_string(),
+                })
+                .collect(),
+            expiration: self.expiration.map(super::convert_stamp),
+        }
+    }
+}
+
+impl StargateMessage for PeriodicAllowance {
+    type ProtoType = feegrant::v1beta1::PeriodicAllowance;
+
+    fn to_proto(&self) -> feegrant::v1beta1::PeriodicAllowance {
+        feegrant::v1beta1::PeriodicAllowance {
+            basic: self.basic.clone().map(|b| b.to_proto()),
+            period: self.period.map(|p| prost_types::Duration {
+                seconds: p.as_secs() as i64,
+                nanos: 0,
+            }),
+            period_spend_limit: super::convert_coins(self.period_spend_limit.clone()),
+            period_can_spend: super::convert_coins(self.period_can_spend.clone()),
+            period_reset: self.period_reset.map(super::convert_stamp),
+        }
+    }
+}
+
+impl<A: BasicOrPeriodicAllowance> StargateMessage for AllowedMsgAllowance<A> {
+    type ProtoType = feegrant::v1beta1::AllowedMsgAllowance;
+
+    fn to_proto(&self) -> feegrant::v1beta1::AllowedMsgAllowance {
+        feegrant::v1beta1::AllowedMsgAllowance {
+            allowance: self.allowance.as_ref().map(|a| a.to_any()),
+            allowed_messages: self.allowed_messages.clone(),
+        }
+    }
+}
